+++ src/route/mod.rs
@@ -1,0 +1,5 @@
+mod r_login;
+mod r_callback;
+
+pub use r_login::login;
+pub use r_callback::callback;
