--- src/util/env_var.rs
@@ -1,3 +1,0 @@
-pub fn env_var(name: &str) -> String {
-    std::env::var(name).expect(&format!("{} not set", name))
-}
