--- src/util/mod.rs
@@ -1,5 +1,0 @@
-mod random_string;
-mod env_var;
-
-pub use random_string::random_string;
-pub use env_var::env_var;
